Notification page&manager

+++ client/src/components/pages/AppointmentManager.js
@@ -267,7 +267,7 @@ function AppointmentManager() {
                       >
                         <Button
                           linkon="0"
-                          onClick={() => {
+                          onClick={() => { if (window.confirm("Sigurno želite obrisati?"))
                             deleteAppointment(row["idorder"]);
                           }}
                           buttonstyle="btn--primary"
